feature / role CRUD

+++ internal/entity/role.go
@@ -1,7 +1,7 @@
 package entity
 
 type Role struct {
-	Id        int    `db:"id"`
+	Id        int64  `db:"id"`
 	Name      string `db:"name"`
 	Authority int32  `db:"authority"`
 }
